app/pyproject.toml
[tool.ruff.lint]
ignore = ["F821"]

[tool.pdm]
distribution = false

[project]
name = "task_tracker"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Виктор Сагалаков", email = "sagalakov.viktor@gmail.com" },
]
dependencies = [
    "bcrypt==4.0.1",
    "passlib>=1.7.4",
    "fastapi>=0.119.0",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "redis>=6.4.0",
    "pyjwt>=2.10.1",
    "pydantic[email]>=2.12.2",
]
requires-python = "==3.13.*"
readme = "README.md"
license = { text = "MIT" }


[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "ruff>=0.14.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.21.0",
]
